[NO GBP] contributor.md web links (#92507)
… link
## About The Pull Request

A very minor update to contributor guidelines - I've been going through
the wiki pages and guides and trying to figure out what would make
contributing a bit more accessible - When you clicked on the .md links
in the guide, it would 404 on Github as it was expecting a local repo
path instead of a web path.

### Steps to re-produce

1. Start here: https://github.com/tgstation/tgstation
2. Scroll down and click Contributing tab button
3. Click under the contents "Development guides"
4. Then click one of the links
5. Currently the URL is:
https://github.com/tgstation/tgstation/blob/master/guides/HARDDELETES.md
6. But the site is not knowing it is actually:
https://github.com/tgstation/tgstation/blob/master/.github/guides/HARDDELETES.md

The .github is missing from the URL.

I also spotted a link in the atmos guide without a title so added it +
fixed the URL.

Originally added double absolute web links but on a comment - Changed it
from **./ to /.**


https://docs.github.com/en/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax#relative-links

Fully unsolved and not acknowledged alas, Quite a few discussions and
sub links here:

https://github.com/orgs/community/discussions/67750

No change log as does not affect anything server side.
## Why It's Good For The Game

Keen to lower the friction barriers for people who might be curious to
contribute and have never done it before.
## Changelog

N/A - No change log as does not affect anything server side.