Mobs with binary channel access now recieve a "New Host" message when a new AI is brought online. (#92334)
## About The Pull Request
Mobs with binary channel access will get a message (and sound alert)
when a new AI is created. The message includes the new AI's name, but no
location or any other information.

AIs spawned by the Syndicate Intellicard item will not send a message.
This also only affects AIs, so Cyborgs and Binary Key users likewise do
not create an alert.

New AIs spawning on Centcom for Admin nonsense also will not send an
alert.
## Why It's Good For The Game
Binary is normally a reasonably safe place for AIs to discuss things
with their borgs, and while it can be tapped into by various other
users, sudden new AIs are the most likely to be detrimental if the older
AI is on a harmful lawset. So now existing AIs get an alert that a new
host is online and, more importantly, listening in.

Syndicate AIs, of course, need to be secret and get a pass on making
alerts.
## Changelog
:cl:
add: AIs now get an alert when a new (non-syndicate) AI is brought
online.
/:cl:

---------

Co-authored-by: Ghom <42542238+Ghommie@users.noreply.github.com>