Add bind (shift-alt-primary) specifically for opening the TGUI loot panel (#92813)
## About The Pull Request

Adds a new bind `AltShiftClickOn` for explicitly opening the TGUI loot
panel at some tile without raising any signals or calling into an
override-able proc. This bind is in addition to the already existing
alt-click and it is not intended to replace it.

This is similar to [this
pr](https://github.com/tgstation/tgstation/pull/92307) but does not
replace the old bind nor add any features beyond the bind opening the
loot panel.

## Why It's Good For The Game

Originally I started looking into this because it was very obnoxious
that I was able to interact with pipes placed in walls, but not pipes
placed underneath windows. I had thought the issue was because the
window overlapped the pipe and was unable to open the loot panel to get
at it because the window handles alt-clicks through the
`simple_rotation` component's signal handler. Unfortunately the pipe is
still obscured by the window, so this does not fix that.

However I still see this as useful as it allows opening the very nice
and wonderful loot panel on a pile of items which may or may not handle
the alt click signal without needing to hunt for a pixel of the tile (or
of an item you _think_ may not handle it) to do so.

## Changelog

:cl:
qol: shift+alt will open the tgui loot panel on any tile without being
intercepted by items that may be on the tile
code: adds a shift+alt click on proc for living mobs to explicitly open
the tgui loot panel in addition to the already existing alt bind
/:cl: