Removes smart tracking from the AI's Crew Monitor (#92977)
## About The Pull Request

Listen dude I just really hate those fucking sandbraine- 
In order to auto-disable tracking if a mob is not well, trackable (read,
wrong z, out of camera sight), we need to well, check if the mob is in
sight of a camera.

This requires cameras to update if they're marked dirty (which they will
be since any opacity change (read: airlocks closing) does this), sooooo
once every 10 seconds when the ui updates we have to update what like 60
camera chunks pessimistically?

That eats a variable amount of cost, somewhere between 300% and 600% of
the tick on catwalk in the round I'm observing, tho it kind of depends
on both the amount of z levels and pop count and so on.

Anyway so instead of doing that let's just not tell them if it's
disabled (like we do for every other button that does this). It's a UX
hit but I cannot come up with an acceptable solution, and it can ALREADY
be wrong because it's behind a timer, so like...

## Why It's Good For The Game

Hopefully significantly less random lagspikes at anything above lowpop

<img width="901" height="29" alt="image"
src="https://github.com/user-attachments/assets/936a6d91-aab9-4686-98e6-cd93b586227c"
/>

<img width="895" height="51" alt="image"
src="https://github.com/user-attachments/assets/bb6782c9-9e67-42dd-a241-cc5bf795744b"
/>

## Changelog
:cl:
balance: AIs viewing the crew monitor will no longer be told ahead of
time if they cannot jump to a player.
/:cl: