[NO GBP] Un-reverts some core crafting changes (with better fix) (#93545)
## About The Pull Request
The soft-revert was merged since it targeted an issue that could crash
the server with infinite loops. However, the solution was kinda shoddy,
which brought the `get_environment` inside the main loop, which I'm once
again trying to move outside it. This way it only has to be called once.

The NO GBP tag is here because the issue was my fault. This PR includes
some small changes done to test and ensure that this won't happen in the
future.

## Why It's Good For The Game
Fewer unnecessary calls. The performance of the crafting unit test is
roughly the same (10 seconds circa), because the bulk of it is spawning
the required objects anyway.

## Changelog
N/A