actually makes it so you can arm guncase evidence disposal bombs (#94015)
## About The Pull Request
About what it says on the tin. I don't know how long it's been broken
for but now you can finally use this as the funny explosive it was
always meant to be.

## Why It's Good For The Game

haha guncase go boom

## Changelog

:cl:
fix: Makarov gun/ammo cases (the ones with the evil red S and the
evidence disposal bomb) no longer require you to be stunned to arm the
evidence disposal bomb.
/:cl:

Co-authored-by: Hatterhat <Hatterhat@users.noreply.github.com>