2 new virus symptoms (#93943)
## About The Pull Request

This PR adds 2 new virus symptoms for virology: Anticoagulant and
Thermoregulation. Yes the names are really long I dunno what else to
call them. Anticoagulant is a stealth oriented level 7 negative virus
symptom that makes the host bleed more and with high enough stats also
increases their susceptibility to recieving bleeding wounds.
Thermoregulation is a level 5 positive virus symptom that regulates your
body temperature based on your species.

Anticoagulant isn't really good for a death virus but the idea is that
it's really stealthy so you can put it in a stealth virus and then grab
a bleeding weapon and do a lot of damage with it.

## Why It's Good For The Game

New content.

## Changelog

:cl:
add: Added 2 new virus symptoms
/:cl:

---------

Co-authored-by: Fghj240 <fakeemail@notrealemail.com>
Co-authored-by: Aki Ito <11748095+ExcessiveUseOfCobblestone@users.noreply.github.com>