Adds jump and search functions to the planecube debugger (#94161)
## About The Pull Request

Adds a dropdown listing all planes to the top bar of the planecube
debugger which sends the view to the selected plane upon using. Double
clicking connection nodes also sends the camera to the plane the other
end of the connection is linked to.

## Why It's Good For The Game

Lets us code and debug stuff easier

## Changelog
:cl:
admin: Added jump and search functions to the planecube debugger
/:cl: