Axes harvest products into procs instead of lists (#94233)
## About The Pull Request

###### Get it? Cause you're chopping down trees...

Why do this at all? Well, we have been looking into memory consumption
and found an unexpectedly high amount of lists on flora objects.

Was talking about it with Melbert, we agreed this would be a good way of
removing needless lists on every blade of grass that are just taking up
memory.

Instead, these lists will be generated and consumed when the actually
chopping down of a tree occurs.

## Why It's Good For The Game

Uploads some more RAM

## Changelog

Not player-facing, things will function the same as they always have.