Plasma is once again not affected by elasticity. (#94394)
## About The Pull Request

With the changes to elasticity in #93235, elasticity was reworked.
However, whether intentional or not, plasma had it's k_elasticity of 0
removed, meaning that plasma sheets were made elastic. Plasma is unique,
in that it's intended to have a static export value grandfathered in
from the old days :tm:, as well as since it's not sold on the GMM,
there's less of a point of having it be affected by elasticity.

This PR Reimplements it's immunity to elasticity.

## Why It's Good For The Game

This is old, old precedent and I think got roped into the elasticity
rework without having the context that it's always been immune to
elasticity.

In addition, it makes sense that the plasma research station would be
the ones setting the price on their unique, rare material. I understand
that cargo isn't hurting for ways to make more money than god right now,
but I'm confident in locking this down as precedent.

## Changelog

:cl:
fix: Plasma sheets are once again immune to market elasticity, selling
for a static 80 credits each yet again.
/:cl: