Rewrites chat settings internals (#94436)
## About The Pull Request
Atomized from a larger PR. There should be no UI behavior changes with
this. It rewrites how settings are handled in tgui-panel internals.
Previously, all UI actions go through TGUI's bespoke redux
implementation. This is all part of the broader system we used to give
us reactive state in inferno. We have the real tools now - we shouldn't
be using this to get by. For the most part, I find
middleware/reducers/actions much more confusing.

Previously:
- user clicks open settings
- this fires a message ('settings/toggle')
- this message is intercepted by the store
- the message is filtered through a sieve of middleware and reducers
- the appropriate action is taken
- rerender

It's a miracle all of this works after having so many cooks in the
kitchen.

Now:
- user clicks open
- it directly flips the settingsVisible state
- rerender

BONUS: This now does runtime type checking, which should fix broken
settings and reduce the need to clear cache
## Why It's Good For The Game
It's done in a React-y way and should be easier to work on. 
Eventually, I want to remove the redux implementation to just work with
the real state tools. This is a huge chunk.
Hopefully more reliable settings panel?
## Changelog

---------