Donk shotgun shell boxes can now hold shotgun shells (#94498)
## About The Pull Request
Shells spawn in the box, but when removed they can't be put back in.
This is because they hold only donk pockets.

Fixes this, and also allows you to refill it with other shotgun shells,
since they're all more or less the same shape. This mirrors how the
other donk boxes can store donk pockets of any kind.
## Why It's Good For The Game
Exit-only boxes for ammo are very frustrating for players. They expect
to be able to put the ammo back inside.
## Changelog
:cl:
fix: The Donk Co. 'Donk Spike' boxes now properly store shotgun shells.
/:cl: