Fixes everything that should care if they don't have enough power not caring if they don't have enough power (#94576)
## About The Pull Request

[cost of operation] < [available consumed power] is a check that will
never succeed, because it'll never have consumed more power than
`amount`. Currently, this doesn't cause any issues when there is enough
power because it's < and not <=. But it does mean that machines that
check if use_energy() succeeded will never fail. Unless there's no APC.

## Why It's Good For The Game

Checks checking for what they're checking for is generally a positive
for the utility of said checks

## Changelog
:cl:

fix: Things that are supposed to be able to fail to draw enough power to
operate are actually able to do so.

/:cl: