Makes the BSA (always) cost power to fire (#94579)
## About The Pull Request

Removes what appears to have been an attempt at making a cooldown from
the BSA's code(it did nothing) and adds a check to right before you fire
that ensures the full cost of the shot has been paid before firing.
Reliant on #94576 merging to do anything.

## Why It's Good For The Game

You can currently fire the bsa so fast the only restriction on firerate
is getting rate limited from clicking FIRE too fast, and it costs jack
shit. You could make it cost even less by making an apc that has a AA
battery for a cell. I think this is stupid, and that the BSA should
require a consistent and gargantuan amount of power to fire. Every shot
should be valuable, and currently, they're not.

## Changelog
:cl:

balance: The BSA now always requires its full power cost paid in order
to fire

/:cl:

---------

Co-authored-by: SyncIt21 <110812394+SyncIt21@users.noreply.github.com>