Changes Coveroli in Albuterol's recipe to standard Convermol (#94578)
## About The Pull Request

Replaces Coveroli (Convermol's inverse) in Albuterol's recipe, and the
reversal recipe, to its base form.
Closes #94574

## Why It's Good For The Game

Despite it not being a bug, LT3 brought up a good point in #94574 about
the requirements being overly harsh. It doesn't make much sense for a
somewhat mediocre drug whose only purpose outside of niche interactions
with nitrium and alike is to cure asthma to require inverse chemicals.
We already have a design document to retire fermichem as-is, and forcing
obscure mechanics onto players as a requirement for rather basic
medicines is not a good idea, especially considering none of our in-game
guides (nor the wiki) explain inverse recipes either.

## Changelog
:cl:
balance: Changed Coveroli in Albuterol's recipe to standard Convermol
/:cl: